pyproject.toml
[tool.poetry]
name = "trail-guide-content-server"
version = "0.17.2"
description = "A server for storing content for a trail guide app."
authors = ["David Lougheed <david.lougheed@gmail.com>"]
license = "GPL-3.0"
readme = "README.md"
packages = [{include = "trail_guide_content_server"}]
include = [
    "trail_guide_content_server/migrations/*.sql",
    "trail_guide_content_server/schema.sql",
]
classifiers=[
    "Programming Language :: Python :: 3 :: Only",
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
    "Operating System :: OS Independent",
]

[tool.poetry.dependencies]
python = "^3.11"
click = "^8.1.7"
Flask = "^3.1.1"
Flask-Cors = "^6.0.1"
jsonschema = "^4.23.0"
PyJWT = {extras = ["crypto"], version = "^2.8.0"}
python-dotenv = "^1.0.1"
qrcode = {extras = ["pil"], version = "^8.2.0"}
requests = "^2.31.0"
Werkzeug = "^3.0.1"
pydantic = "^2.9.0"


[tool.poetry.group.dev.dependencies]
mypy = "~1.16.1"
pytest = "^8.3.2"
types-jsonschema = "~4.23.0.20240712"
types-Flask-Cors = "^4.0.0.20240523"
types-qrcode = "^8.2.0.20250506"
coverage = "^7.6.1"
pytest-cov = "^5.0.0"
ruff = "^0.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 120
